구독상세p/정보 탭 > 거래처 정보 섹션 - 담당자,이메일,전화번호 validation 체크
